# api/wwwroot/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Rates API",
        "license": {
        "name": "MIT"
        }
    },
    "host": "rates.nicoll.io",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json",
        "application/xml"
    ],
    "produces": [
        "application/json",
        "application/xml"
    ],
    "paths": {
        "/rates": {
          "get": {
            "summary": "List all rates",
            "operationId": "listrates",
            "tags": [
              "rates"
            ],
            "parameters": [
              {
                "name": "limit",
                "in": "query",
                "description": "How many items to return at one time (max 100)",
                "required": false,
                "type": "integer",
                "format": "int32"
              }
            ],
            "responses": {
              "200": {
                "description": "An paged array of rates",
                "headers": {
                  "x-next": {
                    "type": "string",
                    "description": "A link to the next page of responses"
                  }
                },
                "schema": {
                  "$ref": "#/definitions/rates"
                }
              },
              "default": {
                "description": "unexpected error",
                "schema": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
          },
          "post": {
            "summary": "Create a rate",
            "operationId": "createrates",
            "tags": [
              "rates"
            ],
            "responses": {
              "201": {
                "description": "Null response"
              },
              "default": {
                "description": "unexpected error",
                "schema": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
          }
        },
        "/rates/{id}": {
          "get": {
            "summary": "Info for a specific rate",
            "operationId": "showrateById",
            "tags": [
              "rates"
            ],
            "parameters": [
              {
                "name": "rateId",
                "in": "path",
                "required": true,
                "description": "The id of the rate to retrieve",
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "Expected response to a valid request",
                "schema": {
                  "$ref": "#/definitions/rates"
                }
              },
              "default": {
                "description": "unexpected error",
                "schema": {
                  "$ref": "#/definitions/Error"
                }
              }
            }
          }
        }
      },
      "definitions": {
        "Pet": {
          "required": [
            "id",
            "name"
          ],
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "name": {
              "type": "string"
            },
            "tag": {
              "type": "string"
            }
          }
        },
        "Pets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Pet"
          }
        },
        "Error": {
          "required": [
            "code",
            "message"
          ],
          "properties": {
            "code": {
              "type": "integer",
              "format": "int32"
            },
            "message": {
              "type": "string"
            }
          }
        }
      }
}